@@ -12,39 +12,42 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "alt": "0.16.10",
-    "highlight.js": "8.8.0",
-    "hjs-webpack": "2.8.1",
-    "lodash": "3.9.3",
+    "alt": "0.17.8",
+    "highlight.js": "8.9.1",
+    "hjs-webpack": "3.0.0",
+    "lodash": "3.10.1",
     "normalize.css": "3.0.3",
-    "object-assign": "3.0.0",
-    "radium": "0.13.2",
-    "react": "0.13.3",
-    "react-router": "1.0.0-beta2",
-    "react-tween-state": "0.0.5",
-    "style-loader": "0.12.3",
-    "stylus-loader": "1.2.1",
-    "tween-functions": "1.0.2"
+    "radium": "0.14.2",
+    "react": "0.14.2",
+    "react-dom": "0.14.2",
+    "react-addons-transition-group": "0.14.2",
+    "history": "1.13.0",
+    "react-router": "1.0.0-rc4",
+    "react-tween-state": "0.1.3",
+    "tween-functions": "1.1.0"
   },
   "devDependencies": {
-    "autoprefixer-core": "^5.2.1",
-    "babel": "^5.6.14",
-    "babel-core": "^5.6.15",
+    "autoprefixer-core": "^6.0.1",
+    "babel": "^5.8.29",
+    "babel-core": "^5.8.30",
     "babel-eslint": "^4.1.3",
-    "babel-loader": "^5.2.2",
-    "css-loader": "^0.15.1",
-    "eslint": "^1.5.1",
-    "eslint-config-defaults": "^6.0.0",
+    "babel-loader": "^5.3.2",
+    "css-loader": "^0.21.0",
+    "eslint": "^1.8.0",
+    "eslint-config-defaults": "^7.1.1",
     "eslint-plugin-filenames": "^0.1.2",
-    "eslint-plugin-react": "^3.4.2",
+    "eslint-plugin-react": "^3.6.3",
     "file-loader": "^0.8.4",
-    "json-loader": "^0.5.2",
-    "node-libs-browser": "^0.5.2",
+    "json-loader": "^0.5.3",
+    "node-libs-browser": "^0.5.3",
+    "postcss-loader": "^0.7.0",
     "raw-loader": "^0.5.1",
-    "react-hot-loader": "^1.2.7",
+    "react-hot-loader": "^1.3.0",
+    "style-loader": "0.13.0",
+    "stylus-loader": "1.4.2",
     "surge": "latest",
     "url-loader": "^0.5.6",
-    "webpack": "^1.10.0",
-    "webpack-dev-server": "^1.10.0"
+    "webpack": "^1.12.2",
+    "webpack-dev-server": "^1.12.1"
   }
 }
