@@ -1,22 +1,21 @@
-import React from "react/addons";
-import assign from "object-assign";
+import React, { Component, PropTypes } from "react";
 import Radium from "radium";
 
 @Radium
-class S extends React.Component {
+class S extends Component {
   render() {
     let styles = {};
     if (this.props.type.indexOf("strikethrough") !== -1) {
-      styles = assign(styles, {textDecoration: "line-through"});
+      styles = Object.assign(styles, {textDecoration: "line-through"});
     }
     if (this.props.type.indexOf("underline") !== -1) {
-      styles = assign(styles, {textDecoration: "underline"});
+      styles = Object.assign(styles, {textDecoration: "underline"});
     }
     if (this.props.type.indexOf("bold") !== -1) {
-      styles = assign(styles, {fontWeight: "bold"});
+      styles = Object.assign(styles, {fontWeight: "bold"});
     }
     if (this.props.type.indexOf("italic") !== -1) {
-      styles = assign(styles, {fontStyle: "italic"});
+      styles = Object.assign(styles, {fontStyle: "italic"});
     }
     return (
       <span style={[styles, this.context.styles.components.s[this.props.type], this.props.style]}>
@@ -27,13 +26,13 @@ class S extends React.Component {
 }
 
 S.propTypes = {
-  children: React.PropTypes.node,
-  style: React.PropTypes.object,
-  type: React.PropTypes.array
+  children: PropTypes.node,
+  style: PropTypes.object,
+  type: PropTypes.array
 };
 
 S.contextTypes = {
-  styles: React.PropTypes.object
+  styles: PropTypes.object
 };
 
 export default S;
