@@ -1,6 +1,6 @@
 /*global window*/
 
-import React from "react/addons";
+import React, { PropTypes } from "react";
 import Base from "./base";
 import Radium from "radium";
 
@@ -28,8 +28,8 @@ class Text extends Base {
   }
   resize() {
     if (this.props.fit) {
-      const text = React.findDOMNode(this.refs.text);
-      const container = React.findDOMNode(this.refs.container);
+      const text = this.refs.text;
+      const container = this.refs.container;
       text.style.display = "inline-block";
       const scale = (container.offsetWidth / text.offsetWidth);
       const height = text.offsetHeight * scale;
@@ -57,21 +57,25 @@ class Text extends Base {
         transformOrigin: "center top"
       }
     };
-    return this.props.fit
-    ? <div
-        ref="container"
-        style={[
-          this.context.styles.components.text,
-          this.getStyles(), styles.container]}>
-        <span
-          ref="text"
-          style={[styles.text, this.props.style]}>
+    return (
+      this.props.fit ? (
+        <div
+          ref="container"
+          style={[this.context.styles.components.text, this.getStyles(), styles.container]}
+        >
+          <span
+            ref="text"
+            style={[styles.text, this.props.style]}
+          >
+            {this.props.children}
+          </span>
+        </div>
+      ) : (
+        <p style={[this.context.styles.components.text, this.getStyles(), this.props.style]}>
           {this.props.children}
-        </span>
-      </div>
-    : <p style={[this.context.styles.components.text, this.getStyles(), this.props.style]}>
-        {this.props.children}
-      </p>;
+        </p>
+      )
+    );
   }
 }
 
@@ -80,12 +84,12 @@ Text.defaultProps = {
 };
 
 Text.propTypes = {
-  children: React.PropTypes.node,
-  lineHeight: React.PropTypes.number
+  children: PropTypes.node,
+  lineHeight: PropTypes.number
 };
 
 Text.contextTypes = {
-  styles: React.PropTypes.object
+  styles: PropTypes.object
 };
 
 export default Text;
